Refactor recovery. Split into files and abstract serializing behaviour.
