+++ nomos-services/utils/src/overwatch/recovery/errors.rs
@@ -1,0 +1,10 @@
+// STD
+use std::io;
+
+#[derive(thiserror::Error, Debug)]
+pub enum RecoveryError {
+    #[error(transparent)]
+    IoError(#[from] io::Error),
+    #[error(transparent)]
+    SerdeError(#[from] serde_json::Error),
+}
